Tests/LibWeb/Text/expected/MathML/MathMLAnchorElement.txt

<a> tag inside <math> is: MathMLAnchorElement
Its href is: http://example.com/
